+++ src/middlewares/guard.js
@@ -6,3 +6,14 @@ export const isAdmin = (req, res, next) => {
       message: 'unauthorized'
     });
 }
+
+export const isLoggedIn = function (req, res, next) {
+  if (req.user) {
+    next();
+  } else {
+    res.status(401).send({
+      status: 'error',
+      message: 'unauthorized to chat without logging in'
+    });
+  }
+}